Refactor git package into atomic files
